IPEDS (Integrated Postsecondary Education Data System) data, 2017

A dataset formed by extracting information from the surveys conducted by the Institute of Education Sciences in 2017 (link provided as a source). These surveys are conducted annually by NCES () to gather information from every postsecondary institution that participates in federal student financial aid programs.

Usage

ipeds

Format

A data frame with 6,440 rows and 60 columns. Each row corresponds to an institution.

unitid

(character) unique ID number assigned to postsecondary institutions surveyed through IPEDS

instnm

(character) institution name

state

(factor) US Postal Service state abbreviation

addr

(character) street address or post office box

zip

(factor) 5-digit ZIP code

gentele

(character) ten-digit telephone number

webaddr

(character) Internet website address

longitud

(double) longitude of location

latitude

(double) latitude of location

urbanicity

(factor) whether the location of the institution is rural or in a town, city, or suburb

degree_level

(factor) highest level of degree offered; 1 = four or more years, 2 = at least two but less than four years, 3 = less than two years

control_level

(factor) whether the institution is publicly or privately run; 1 = public, 2 = private and not-for-profit, 3 = private and for-profit

graduate

(factor) whether the institution offers graduate degrees or certificates; 0 - no, 1 = yes

hs_remedial_services

(factor) whether the institution offers adult basic or remedial instruction or high school equivalency courses; 0 = no, 1 = yes

remedial_services

(factor) whether the institution offers remedial services -- instructional activities designed for students deficient in the general competencies necessary for a regular postsecondary curriculum and educational setting; 0 = no, 1 = yes

multisystem_status

(factor) whether the institution is part of a multi-institution or multi-campus organization that owns, governs, or controls the institution. A multi-institution or multi-campus organization includes organizations with two or more institutions or campuses. Non-postsecondary education agencies that govern or control institutions include, but are not limited to, public school districts, art organizations, hospitals and other medical/health organizations. The following are not included: coordinating systems, single institution owner, single institution corporate name, single institution governing board, consortia, associations, or religious affiliations. 0 = no, 1 = yes

parentsystem

(character) NA if multisystem_status is 0; otherwise, name of the multi-institution or multi-campus organization

pct_female

(double) proportion of women enrolled for credit during the 12-month reporting period

pct_nat_aa

(double) proportion of American Indian or Alaska Native men and women enrolled for credit during the 12-month period.

American Indian or Alaska Native - A person having origins in any of the original peoples of North and South America (including Central America) who maintains cultural identification through tribal affiliation or community attachment.

pct_asian

(double) proportion of Asian men and women enrolled for credit during the 12-month period.

Asian - A person having origins in any of the original peoples of the Far East, Southeast Asia, or the Indian Subcontinent, including, for example, Cambodia, China, India, Japan, Korea, Malaysia, Pakistan, the Philippine Islands, Thailand, and Vietnam.

pct_black

(double) proportion of Black or African American men and women enrolled for credit during the 12-month period.

Black or African American - A person having origins in any of the black racial groups of Africa.

pct_hispanic

(double) proportion of Hispanic or Latino men and women enrolled for credit during the 12-month period.

Hispanic or Latino - A person of Cuban, Mexican, Puerto Rican, South or Central American, or other Spanish culture or origin, regardless of race.

pct_nat_hpi

(double) proportion of Native Hawaiian or Other Pacific Islanders men and women enrolled for credit during the 12-month period.

Native Hawaiian or Other Pacific Islanders - A person having origins in any of the original peoples of Hawaii, Guam, Samoa, or other Pacific Islands.

pct_white

(double) proportion of White men and women enrolled for credit during the 12-month period.

White - A person having origins in any of the original peoples of Europe, the Middle East, or North Africa.

pct_2more

(double) proportion of men and women of two or more races enrolled for credit during the 12-month period.

Two or more races - Category used by institutions to report persons who selected more than one race.

pct_unknown

(double) proportion of race/ethnicity unknown men and women enrolled for credit during the 12-month period.

This category is used ONLY if the student did not select a racial/ethnic designation, AND the postsecondary institution finds it impossible to place the student in one of the aforementioned racial/ethnic categories during established enrollment procedures or in any post-enrollment identification or verification process.

pct_foreign

(double) proportion of nonresident alien men and women enrolled for credit during the 12-month reporting period.

Nonresident alien - A person who is not a citizen or national of the United States and who is in this country on a visa or temporary basis and does not have the right to remain indefinitely.

NOTE - Nonresident aliens are included here, rather than in any of the five racial/ethnic categories. Resident aliens and other eligible (for financial aid purposes) non-citizens who are not citizens or nationals of the United States and who have been admitted as legal immigrants for the purpose of obtaining permanent resident alien status (and who hold either an alien registration card (Form I-551 or I-151), a Temporary Resident Card (Form I-688), or an Arrival-Departure Record (Form I-94) with a notation that conveys legal immigrant status such as Section 207 Refugee, Section 208 Asylee, Conditional Entrant Parolee or Cuban-Haitian) are to be reported in the appropriate racial/ethnic categories along with United States citizens.

pct_over25_total

(double) proportion of full time students over the age of 25 enrolled for credit

pct_parttime

(double) proportion of part time students enrolled for credit

admit_rate

(double) proportion of first-time, degree/certificate-seeking undergraduate students who applied, were admitted, and enrolled (full or part time) for the most recent fall period available, out of the total number of applicants

sat

(double) SAT I 50th percentile score for the institution (across verbal and math). Provided ONLY IF test scores are required for admission for first-time, degree/certificate-seeking undergraduate students. (This is applicable to institutions that have no open admission policy for entering first-time undergraduate students).

SAT (Scholastic Aptitude Test) - An examination administered by the Educational Testing Service and used to predict the facility with which an individual will progress in learning college-level academic subjects.

act

(double) ACT Composite 50th percentile score for the institution. Provided ONLY IF test scores are required for admission for first-time, degree/certificate-seeking undergraduate students. (This is applicable to institutions that have no open admission policy for entering first-time undergraduate students).

ACT (American College Testing Program) - The ACT measures educational development and readiness to pursue college-level coursework in English, mathematics, natural science, and social studies.

total_undergrad_lg

(double) grand total number of men and women enrolled for credit during the 12-month reporting period. Log-transformed to reduce extreme positive skew.

title_iv_status

(factor) whether the institution has a written agreement with the Secretary of Education that allows the institution to participate in any of the Title IV federal student financial assistance programs (other than the State Student Incentive Grant (SSIG) and the National Early Intervention Scholarship and Partnership (NEISP) programs). 0 if no, 1 if yes

pct_grant_aid

(double) proportion of undergraduate students awarded grant aid.

Grant aid includes any grant or scholarship aid awarded, from the federal government, a state or local government, the institution, and other sources known by the institution.

pct_pell

(double) proportion of undergraduate students awarded Pell grants.

Pell Grant program (Higher Education Act of 1965, Title IV, Part A, Subpart I, as amended) provides grant assistance to eligible undergraduate postsecondary students with demonstrated financial need to help meet education expenses.

pct_fed_loans

(double) proportion of undergraduate students awarded federal student loans.

Loans to students - Any monies that must be repaid to the lending institution for which the student is the designated borrower. Includes all Title IV subsidized and unsubsidized loans. Does not include PLUS and other loans made directly to parents.

pct_any_aid

(double) proportion of all full-time, first-time degree/certificate-seeking undergraduate students awarded any financial aid.

Financial aid - Grants, loans, assistantships, scholarships, fellowships, tuition waivers, tuition discounts, veteran's benefits, employer aid (tuition reimbursement) and other monies (other than from relatives/friends) provided to students to meet expenses. This includes Title IV subsidized and unsubsidized loans made directly to students.

pct_low_inc

(double) percentage of students in income level 0-30,000

pct_low_mid_inc

(double) percentage of students in income level 30,001-48,000

pct_mid_inc

(double) percentage of students in income level 48,001-75,000

pct_mid_high_inc

(double) percentage of students in income level 75,001-110,000

pct_high_inc

(double) percentage of students in income level 110,001 or more

public

(factor) from control_level; 1 for public institutions, 0 otherwise

private_profit

(factor) from control_level; 1 for private for-profit institutions, 0 otherwise

private_notforprofit

(factor) from control_level; 1 for private not-for-profit institutions, 0 otherwise

city

(factor) from urbanicity; 1 for institutions in cities, 0 otherwise

suburb

(factor) from urbanicity; 1 for institutions in suburbs, 0 otherwise

town

(factor) from urbanicity; 1 for institutions in towns, 0 otherwise

rural

(factor) from urbanicity; 1 for institutions in rural areas, 0 otherwise

Source

https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx

Details

This dataset represents a population of both degree-granting and non-degree-granting active post-secondary institutions of all levels and sectors. It includes public, private for-profit, and private not-for-profit institutions. The vast majority of institutions in the dataset are Title IV institutions (which participate in federal financial aid programs), although some non-Title IV institutions participate as well.