Recruit Units from a Population for Sampling

This function works with the output of 'stratify()'. The user provides the number of units they wish to sample from their population dataset. The function tells the user how many observations to sample from each stratum and generates recruitment lists, one per stratum, which can either be saved to .csv files in any given directory or accessed later on.

Usage

recruit(
  stratify_output,
  guided = TRUE,
  sample_size = NULL,
  save_as_csv = FALSE
)

Arguments

stratify_output

output from 'stratify()', of S3 class 'generalizeR_stratify'

guided

logical; defaults to TRUE. Whether the function should be guided (ask questions and behave interactively throughout) or not. If set to FALSE, must provide values for other arguments below

sample_size

defaults to NULL. If guided is set to FALSE, must provide a number of units to sample

save_as_csv

defaults to NULL. If guided is set to FALSE, specify whether or not to save recruitment lists to working directory; TRUE or FALSE

Value

A three-element list containing the recruitment lists and the recruitment table/kable

Details

This function, and the others in this package, are designed to mimic the website https://www.thegeneralizer.org/ based on the papers referenced below.

References

Tipton, E. (2014). Stratified sampling using cluster analysis: A sample selection strategy for improved generalizations from experiments. Evaluation Review, 37(2), 109-139.

Tipton, E. (2014). How generalizable is your experiment? An index for comparing experimental samples and populations. Journal of Educational and Behavioral Statistics, 39(6), 478-501.

Examples

# \donttest{
if (FALSE) {
# Guided:
recruit(ipeds_stratify_output_obj)

# Not guided:
recruit(ipeds_stratify_output_obj, guided = FALSE, sample_size = 72, save_as_csv = TRUE)
}
# }